fix: add option to not exclude imports in loadDocument

+++ packages/language/src/document.ts
@@ -32,8 +32,10 @@ import type { ZModelFormatter } from './zmodel-formatter';
 export async function loadDocument(
     fileName: string,
     additionalModelFiles: string[] = [],
+    keepImports: boolean = false,
 ): Promise<
-    { success: true; model: Model; warnings: string[], services: ZModelServices } | { success: false; errors: string[]; warnings: string[] }
+    | { success: true; model: Model; warnings: string[]; services: ZModelServices }
+    | { success: false; errors: string[]; warnings: string[] }
 > {
     const { ZModelLanguage: services } = createZModelServices(false);
     const extensions = services.LanguageMetaData.fileExtensions;
@@ -121,14 +123,16 @@ export async function loadDocument(
 
     const model = document.parseResult.value as Model;
 
-    // merge all declarations into the main document
-    const imported = mergeImportsDeclarations(langiumDocuments, model);
+    if (keepImports === false) {
+        // merge all declarations into the main document
+        const imported = mergeImportsDeclarations(langiumDocuments, model);
 
-    // remove imported documents
-    imported.forEach((model) => {
-        langiumDocuments.deleteDocument(model.$document!.uri);
-        services.shared.workspace.IndexManager.remove(model.$document!.uri);
-    });
+        // remove imported documents
+        imported.forEach((model) => {
+            langiumDocuments.deleteDocument(model.$document!.uri);
+            services.shared.workspace.IndexManager.remove(model.$document!.uri);
+        });
+    }
 
     // extra validation after merging imported declarations
     const additionalErrors = validationAfterImportMerge(model);
